adds CardinalityEstimation as an interface
